Fix devilish bug involving library loading
The problem was by theme didn't correctly specify its dependencies, that is
dash.el. The only reason it had been working this long was because
Smartparens (which also loads Dash) had always been loaded before it. But not
because my Smartparens config came above it in the file (it didn't, and that
wouldn't have loaded it immediately anyway). Because there was a line added by
Customize at some point in temp-custom.el, `(smartparens-global-mode t)`. And
because I happend to load temp-custom.el before my Appearance config (and
generally close to the top of my config). So Smartparens got autoloaded without
any intention on my part.

Not fun to debug.